+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Switch to Buildkite and add test analytics
+
 ## v0.3.6
 
 - Add image_draw!