+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Switch to Buildkite and add test analytics
 - Switch Taylor CLI to BuildKite and add test analytics
+- Fix up the README
 
 ## v0.3.6